separate extensions/v1alpha1/helper tests
